+++ CHANGELOG.md
@@ -33,6 +33,7 @@ for now removed features.
 - **grid**: fix a behavior where tabbing to the single-selection radio or clicking on it, inside a grid in a scrollable container, could cause the window to scroll unexpectedly. The window scroll position is now preserved.
 - **radio**: fix a behavior where clicking on a radio button inside a scrollable container would cause the window to scroll to the bottom of the page. Now, the window scroll position is preserved when clicking on a radio button.
 - **radio-group**: fix a11y issues with the hierarchy announcement of radio options by screen readers
+- **input**: remove `aria-labelledby` attribute setting to avoid conflicts with `wcs-form-field` label association
 
 ### Security